resolve the compatibility issue with the newest paddle

+++ examples/IMPALA/train.py
@@ -121,7 +121,7 @@ class Learner(object):
 
                 yield [
                     obs_np, actions_np, behaviour_logits_np, rewards_np,
-                    dones_np, self.lr, self.entropy_coeff
+                    dones_np, np.float32(self.lr), np.float32(self.entropy_coeff)
                 ]
 
     def run_learn(self):
